[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.04 ms, Memory: 75.7 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,9 +1,9 @@
 class Solution {
-    public long[] solution(int x, int n) {
+    public long[] solution(long x, int n) {
         long[] arr = new long[n];
 
         for (int i = 1; i <= n; i++) {
-            arr[i-1] = (long) x * i;
+            arr[i-1] = x * i;
         }
 
         return arr;
